Create README.md
Hi, I’m Michael Smith. I’m interested in finding a great job that is based on my Performance, Effectiveness, Time-Management, and Overall Performance to my Employer.
I'm Interested in What I've done for 20 years.  It doesn't need to Management as before, I am willing to prove myself to the Company who gives me this opportunity.  I'm am Extremely Loyal to them, because if they hire me.  I feel it would be wrong to go to another job over money or whatever the reason is.  Merchandising is more than a job to me, it's interacting with new people and customers.  And going to different stores to either do a reset, planagram, checking dates or placing orders, traveling is what  I love and happens to  bring out the best in me.  It's rare to find a job that doesn't make bored, or wish u didn't have to go work that day.  I'm blessed I found something that I can do as my Career and Enjoy doing the job. 
    I'm looking to Collaborate to make sure myself and hiring Company are on the same page.  Getting off on the right foot is so Important to both parties.  So Im hoping this can happen.  My Email is mikesmith8861@gmail.com and Phone Number is (304-785-4163).
I am having difficulty with my phone saying HTTPS is not allowing me to see what is sent or Documents that won't open.  If this can be looked at I would greatly appreciate it.
Thank you
Michael SMITH